requirements.txt
asttokens==2.0.7
backcall==0.2.0
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
entrypoints==0.4
executing==0.9.1
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.11.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywin32==304
pyzmq==23.2.1
six==1.16.0
stack-data==0.3.0
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
